score updates correctly

+++ Pacman/PacmanEngine/Pacman.h
@@ -18,7 +18,7 @@ namespace pac
 		decltype(GameplaySettings::mPacmanTicksPerMove) GetTicksPerMove() const;
 		void IncreaseScoreCoinCell();
 		void IncreaseScorePowerUpCell();
-		uint16_t GetScore() const;
+		uint64_t GetScore() const;
 		void SetIsPowerUpActive(bool isPowerUpActive);
 		uint16_t GetPowerUpDuration() const;
 		void Draw(IWindow* window) const;
@@ -30,7 +30,7 @@ namespace pac
 		Direction mNextDirection;
 		decltype(GameplaySettings::mPacmanTicksPerMove) mTicksSinceLastMove;
 		decltype(GameplaySettings::mPacmanTicksPerMove) mTicksPerMove;
-		uint16_t mScore;
+		uint64_t mScore;
 		bool mIsPowerUpActive;
 		const uint16_t mPowerUpDuration = 100;
 	};
